Integrate ML cut systematics as Analyzer
A systematic class also derives from Analyzer as it can be treated in
the very same way.

The Analyzer class derives from WorkflowBase as this is going to be a
potential base class for all classes handling/manipulating/deriving
data.

Possibility to implement an AnalyzerAfterBurner to be run after
period-by-period runs have been finished to allow e.g. for merging data
or to apply further analysis steps only possible once all periods have
been processed.

Please also refer to the README.
