+++ machine_learning_hep/analysis/analyzer.py
@@ -12,130 +12,18 @@
 ##   along with this program. if not, see <https://www.gnu.org/licenses/>. ##
 #############################################################################
 
-"""
-main script for doing final stage analysis
-"""
-# pylint: disable=too-many-lines
-import os
-import array
-# pylint: disable=unused-wildcard-import, wildcard-import
-import numpy as np
-# pylint: disable=import-error, no-name-in-module, unused-import
-from root_numpy import hist2array, array2hist
-from ROOT import TFile, TH1F, TH2F, TCanvas, TPad, TF1, TH1D
-from ROOT import gStyle, TLegend, TLine, TText, TPaveText, TArrow
-from ROOT import gROOT, TDirectory, TPaveLabel
-from ROOT import TStyle, kBlue, kGreen, kBlack, kRed, kOrange
-from ROOT import TLatex
-from ROOT import gInterpreter, gPad
 # HF specific imports
-from machine_learning_hep.fitting.helpers import MLFitter
-from machine_learning_hep.logger import get_logger
-from machine_learning_hep.io import dump_yaml_from_dict
-from machine_learning_hep.utilities import folding, get_bins, make_latex_table, parallelizer
-from machine_learning_hep.utilities_plot import plot_histograms
-#from ROOT import RooUnfoldResponse
-#from ROOT import RooUnfold
-#from ROOT import RooUnfoldBayes
-# pylint: disable=too-few-public-methods, too-many-instance-attributes, too-many-statements, fixme
-class Analyzer:
-    species = "analyzer"
-    def __init__(self, datap, case, typean,
-                 resultsdata, resultsmc, valdata, valmc):
-
-        self.logger = get_logger()
-        #namefiles pkl
-        self.case = case
-        self.typean = typean
-        self.datap = datap
-        self.v_var_binning = datap["var_binning"]
-        self.lpt_finbinmin = datap["analysis"][self.typean]["sel_an_binmin"]
-        self.lpt_finbinmax = datap["analysis"][self.typean]["sel_an_binmax"]
-        self.bin_matching = datap["analysis"][self.typean]["binning_matching"]
-        self.p_nptbins = len(self.lpt_finbinmin)
-        analysis_bin_lims_temp = self.lpt_finbinmin.copy()
-        analysis_bin_lims_temp.append(self.lpt_finbinmax[self.p_nptbins-1])
-        self.analysis_bin_lims = array.array('f', analysis_bin_lims_temp)
-        self.lpt_probcutfin = datap["mlapplication"]["probcutoptimal"]
-
-        self.triggerbit = datap["analysis"][self.typean]["triggerbit"]
-
-        self.d_resultsallpmc = resultsmc
-        self.d_resultsallpdata = resultsdata
-
-        n_filemass_name = datap["files_names"]["histofilename"]
-        self.n_filemass = os.path.join(self.d_resultsallpdata, n_filemass_name)
-        self.n_filemass_mc = os.path.join(self.d_resultsallpmc, n_filemass_name)
-        self.n_filecross = datap["files_names"]["crossfilename"]
-        self.p_mass_fit_lim = datap["analysis"][self.typean]['mass_fit_lim']
-        self.p_bin_width = datap["analysis"][self.typean]['bin_width']
-        self.p_num_bins = int(round((self.p_mass_fit_lim[1] - self.p_mass_fit_lim[0]) / \
-                                    self.p_bin_width))
-
-        # Output directories and filenames
-        self.yields_filename = "yields"
-        self.fits_dirname = "fits"
-        self.yields_syst_filename = "yields_syst"
-        self.efficiency_filename = "efficiencies"
-
-        self.n_fileff = datap["files_names"]["efffilename"]
-        self.n_fileff = os.path.join(self.d_resultsallpmc, self.n_fileff)
-        self.n_evtvalroot = datap["files_names"]["namefile_evtvalroot"]
-
-        self.p_indexhpt = datap["analysis"]["indexhptspectrum"]
-        self.p_fd_method = datap["analysis"]["fd_method"]
-        self.p_cctype = datap["analysis"]["cctype"]
-        self.p_sigmav0 = datap["analysis"]["sigmav0"]
-        self.p_inputfonllpred = datap["analysis"]["inputfonllpred"]
-        self.p_latexnmeson = datap["analysis"]["latexnamemeson"]
-
-        self.root_objects = []
-
-        # Fitting
-        self.fitter = None
-
-    @staticmethod
-    def loadstyle():
-        gStyle.SetOptStat(0)
-        gStyle.SetOptStat(0000)
-        gStyle.SetPalette(1)
-        gStyle.SetCanvasColor(0)
-        gStyle.SetFrameFillColor(0)
-
-    @staticmethod
-    def make_pre_suffix(args):
-        """
-        Construct a common file suffix from args
-        """
-        try:
-            _ = iter(args)
-        except TypeError:
-            args = [args]
-        else:
-            if isinstance(args, str):
-                args = [args]
-        args = [str(a) for a in args]
-        return "_".join(args)
-
-    @staticmethod
-    def make_file_path(directory, filename, extension, prefix=None, suffix=None):
-        if prefix is not None:
-            filename = Analyzer.make_pre_suffix(prefix) + "_" + filename
-        if suffix is not None:
-            filename = filename + "_" + Analyzer.make_pre_suffix(suffix)
-        extension = extension.replace(".", "")
-        return os.path.join(directory, filename + "." + extension)
+from machine_learning_hep.workflow.workflow_base import WorkflowBase
 
 
-    def analysis_step(self, step: str):
-        """
-        Given an analysis steps as string, find the corresponding method and call it.
-        Args:
-            step: analysis step as string
-        Returns:
-            True if the analysis step was found and executed, False otherwise
-        """
-        if not hasattr(self, step):
-            return False
-        getattr(self, step)()
-        return True
+class Analyzer(WorkflowBase):
+    def __init__(self, datap, case, typean, period):
+        super().__init__(datap, case, typean, period)
+
+
+
+class AnalyzerAfterBurner(WorkflowBase):
+    def __init__(self, datap, case, typean):
+        super().__init__(datap, case, typean, None)
+
+        self.analyzers = None
