+++ machine_learning_hep/analysis/analyzer_Dhadrons_mult.py
@@ -18,8 +18,7 @@ main script for doing final stage analysis
 # pylint: disable=too-many-lines
 import os
 # pylint: disable=unused-wildcard-import, wildcard-import
-from array import *
-import numpy as np
+from array import array
 # pylint: disable=import-error, no-name-in-module, unused-import
 from root_numpy import hist2array, array2hist
 from ROOT import TFile, TH1F, TH2F, TCanvas, TPad, TF1, TH1D
@@ -35,21 +34,13 @@ from machine_learning_hep.io import dump_yaml_from_dict
 from machine_learning_hep.utilities import folding, get_bins, make_latex_table, parallelizer
 from machine_learning_hep.utilities_plot import plot_histograms
 from machine_learning_hep.analysis.analyzer import Analyzer
-#from ROOT import RooUnfoldResponse
-#from ROOT import RooUnfold
-#from ROOT import RooUnfoldBayes
 # pylint: disable=too-few-public-methods, too-many-instance-attributes, too-many-statements, fixme
 class AnalyzerDhadrons_mult(Analyzer):
     species = "analyzer"
-    def __init__(self, datap, case, typean,
-                 resultsdata, resultsmc, valdata, valmc):
-        super().__init__(datap, case, typean,
-                 resultsdata, resultsmc, valdata, valmc)
-        self.logger = get_logger()
+    def __init__(self, datap, case, typean, period):
+        super().__init__(datap, case, typean, period)
+
         #namefiles pkl
-        self.case = case
-        self.typean = typean
-        self.datap = datap
         self.v_var_binning = datap["var_binning"]
         self.lpt_finbinmin = datap["analysis"][self.typean]["sel_an_binmin"]
         self.lpt_finbinmax = datap["analysis"][self.typean]["sel_an_binmax"]
@@ -64,8 +55,11 @@ class AnalyzerDhadrons_mult(Analyzer):
         self.triggerbit = datap["analysis"][self.typean]["triggerbit"]
         self.p_nbin2 = len(self.lvar2_binmin)
 
-        self.d_resultsallpmc = resultsmc
-        self.d_resultsallpdata = resultsdata
+        self.d_resultsallpmc = datap["analysis"][typean]["mc"]["results"][period] \
+                if period is not None else datap["analysis"][typean]["mc"]["resultsallp"]
+        self.d_resultsallpdata = datap["analysis"][typean]["data"]["results"][period] \
+                if period is not None else datap["analysis"][typean]["data"]["resultsallp"]
+
         self.p_corrmb_typean = datap["analysis"][self.typean]["corresp_mb_typean"]
         if self.p_corrmb_typean is not None:
             self.results_mb = datap["analysis"][self.p_corrmb_typean]["data"]["resultsallp"]
@@ -158,8 +152,10 @@ class AnalyzerDhadrons_mult(Analyzer):
         self.p_sigmamb = datap["ml"]["opt"]["sigma_MB"]
         self.p_br = datap["ml"]["opt"]["BR"]
 
-        self.d_valevtdata = valdata
-        self.d_valevtmc = valmc
+        self.d_valevtdata = datap["validation"]["data"]["dir"][period] \
+                if period is not None else datap["validation"]["data"]["dirmerged"]
+        self.d_valevtmc = datap["validation"]["mc"]["dir"][period] \
+                if period is not None else datap["validation"]["mc"]["dirmerged"]
 
         self.f_evtvaldata = os.path.join(self.d_valevtdata, self.n_evtvalroot)
         self.f_evtvalmc = os.path.join(self.d_valevtmc, self.n_evtvalroot)
@@ -188,37 +184,6 @@ class AnalyzerDhadrons_mult(Analyzer):
         # Fitting
         self.fitter = None
 
-    @staticmethod
-    def loadstyle():
-        gStyle.SetOptStat(0)
-        gStyle.SetOptStat(0000)
-        gStyle.SetPalette(1)
-        gStyle.SetCanvasColor(0)
-        gStyle.SetFrameFillColor(0)
-
-    @staticmethod
-    def make_pre_suffix(args):
-        """
-        Construct a common file suffix from args
-        """
-        try:
-            _ = iter(args)
-        except TypeError:
-            args = [args]
-        else:
-            if isinstance(args, str):
-                args = [args]
-        args = [str(a) for a in args]
-        return "_".join(args)
-
-    @staticmethod
-    def make_file_path(directory, filename, extension, prefix=None, suffix=None):
-        if prefix is not None:
-            filename = Analyzer.make_pre_suffix(prefix) + "_" + filename
-        if suffix is not None:
-            filename = filename + "_" + Analyzer.make_pre_suffix(suffix)
-        extension = extension.replace(".", "")
-        return os.path.join(directory, filename + "." + extension)
 
     # pylint: disable=import-outside-toplevel
     def fit(self):
