--- machine_learning_hep/analysis/systematics_Dhadrons_mult.py
@@ -1,79 +1,0 @@
-#############################################################################
-##  © Copyright CERN 2018. All rights not expressly granted are reserved.  ##
-##                 Author: Gian.Michele.Innocenti@cern.ch                  ##
-## This program is free software: you can redistribute it and/or modify it ##
-##  under the terms of the GNU General Public License as published by the  ##
-## Free Software Foundation, either version 3 of the License, or (at your  ##
-## option) any later version. This program is distributed in the hope that ##
-##  it will be useful, but WITHOUT ANY WARRANTY; without even the implied  ##
-##     warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.    ##
-##           See the GNU General Public License for more details.          ##
-##    You should have received a copy of the GNU General Public License    ##
-##   along with this program. if not, see <https://www.gnu.org/licenses/>. ##
-#############################################################################
-
-"""
-main script for doing final stage analysis
-"""
-# pylint: disable=too-many-lines
-import os
-# pylint: disable=unused-wildcard-import, wildcard-import
-#from array import *
-import numpy as np
-# pylint: disable=import-error, no-name-in-module, unused-import
-from root_numpy import hist2array, array2hist
-from ROOT import TFile, TH1F, TH2F, TCanvas, TPad, TF1, TH1D
-from ROOT import AliHFInvMassFitter, AliVertexingHFUtils, AliHFInvMassMultiTrialFit
-from ROOT import gStyle, TLegend, TLine, TText, TPaveText, TArrow
-from ROOT import gROOT, TDirectory, TPaveLabel
-from ROOT import TStyle, kBlue, kGreen, kBlack, kRed, kOrange
-from ROOT import TLatex
-from ROOT import gInterpreter, gPad
-# HF specific imports
-from machine_learning_hep.fitting.helpers import MLFitter
-from machine_learning_hep.logger import get_logger
-from machine_learning_hep.io import dump_yaml_from_dict
-from machine_learning_hep.utilities import folding, get_bins, make_latex_table, parallelizer
-from machine_learning_hep.utilities_plot import plot_histograms
-from machine_learning_hep.analysis.systematics import Systematics
-
-#from ROOT import RooUnfoldResponse
-#from ROOT import RooUnfold
-#from ROOT import RooUnfoldBayes
-# pylint: disable=too-few-public-methods, too-many-instance-attributes, too-many-statements, fixme
-class SystematicsDhadrons_mult(Systematics):
-    species = "systematics"
-    def __init__(self, datap, case, typean,
-                 resultsdata, resultsmc, valdata, valmc):
-        super().__init__(datap, case, typean,
-                 resultsdata, resultsmc, valdata, valmc)
-        self.logger = get_logger()
-
-        self.p_sgnfunc = datap["analysis"][self.typean]["sgnfunc"]
-        self.p_bkgfunc = datap["analysis"][self.typean]["bkgfunc"]
-        self.p_masspeak = datap["analysis"][self.typean]["masspeak"]
-        self.p_massmin = datap["analysis"][self.typean]["massmin"]
-        self.p_massmax = datap["analysis"][self.typean]["massmax"]
-        self.p_rebin = datap["analysis"][self.typean]["rebin"]
-        self.p_fix_mean = datap["analysis"][self.typean]["fix_mean"]
-        self.p_fix_sigma = datap["analysis"][self.typean]["fix_sigma"]
-
-        self.p_masspeaksec = None
-        self.p_fix_sigmasec = None
-        self.p_sigmaarraysec = None
-        if self.p_sgnfunc[0] == 1:
-            self.p_masspeaksec = datap["analysis"][self.typean]["masspeaksec"]
-            self.p_fix_sigmasec = datap["analysis"][self.typean]["fix_sigmasec"]
-            self.p_sigmaarraysec = datap["analysis"][self.typean]["sigmaarraysec"]
-
-        self.fitter = None
-        self.lpt_finbinmin = datap["analysis"][self.typean]["sel_an_binmin"]
-        self.lpt_finbinmax = datap["analysis"][self.typean]["sel_an_binmax"]
-        self.p_nptfinbins = len(self.lpt_finbinmin)
-
-    # pylint: disable=import-outside-toplevel
-    def cutvariation(self):
-        # Enable ROOT batch mode and reset in the end
-        tmp_is_root_batch = gROOT.IsBatch()
-        gROOT.SetBatch(True)
-        print("Doing cut variation for Dhadrons")
